+++ bot/error.py
@@ -6,13 +6,17 @@ class Error(Exception):
     __ERR_MSG_LIST = {
         0: "Something unexpected happened into the code, please contact the developer",
         1: "TG_BOT_TOKEN Environment have Invalid Bot Token, Please provide the correct token and try again",
-        2: "TG_BOT_TOKEN Environment can't be empty"
+        2: "TG_BOT_TOKEN Environment can't be empty",
+        3: "PORT Environment can't be empty",
+        4: "PORT Environment variable contains invalid port number, Please provide a correct port number and try again"
     }
 
     __ERR_CODE_LIST = {
         "UNEXPECTED_EXCEPTION": 0,
         "INVALID_TG_BOT_TOKEN": 1,
-        "NO_TG_BOT_TOKEN": 2
+        "NO_TG_BOT_TOKEN": 2,
+        "NO_WEB_APP_PORT": 3,
+        "INVALID_WEB_APP_PORT": 4
     }
 
     def __init__(self, ERR_CODE: str, MSG=None):