+++ bot/load_environ.py
@@ -15,13 +15,13 @@ class Environment:
 
         # Verifying
         self.__validate_bot_token()
+        self.__validate_web_app_port()
 
 
     def __load_environ(self):
         self.TG_BOT_TOKEN = os.getenv("TG_BOT_TOKEN")
         self.SPAMWATCH_API_KEY = os.getenv("SPAMWATCH_API_KEY")
-        self.OWNER_NAME = os.getenv("OWNER_NAME")
-        self.OWNER_USERNAME = os.getenv("OWNER_USERNAME")
+        self.WEB_APP_PORT = os.getenv("PORT")
         
     def __validate_bot_token(self):
         if self.TG_BOT_TOKEN is None or len(self.TG_BOT_TOKEN) == 0:
@@ -35,3 +35,16 @@ class Environment:
         
         if not self.TG_BOT_TOKEN[self.TG_BOT_TOKEN.find(":")+1:].isalnum():
             raise error.Error('INVALID_TG_BOT_TOKEN')
+        
+    def __validate_web_app_port(self):
+        if self.WEB_APP_PORT is None or len(self.WEB_APP_PORT) == 0:
+            raise error.Error('NO_WEB_APP_PORT')
+        
+        if not self.WEB_APP_PORT.isdigit():
+            raise error.Error('INVALID_WEB_APP_PORT')
+        
+        if int(self.WEB_APP_PORT) > 65535:
+            raise error.Error('INVALID_WEB_APP_PORT')
+        
+        if int(self.WEB_APP_PORT) < 1:
+            raise error.Error('INVALID_WEB_APP_PORT')