+++ main.py
@@ -1,21 +1,21 @@
 # Web App
 from flask import Flask
 from multiprocessing import Process
+from waitress import serve
+from bot.load_environ import Environment
 
 app = Flask(__name__)
 
-import logging
-logging.getLogger('werkzeug').setLevel(logging.ERROR)
-
 @app.route('/')
 def start():
-    return ""
+    return "Bot is Running"
 
 def run_flask():
-    app.run(host='0.0.0.0', port='5000')
+    serve(app, host='0.0.0.0', port=Environment().WEB_APP_PORT)
 
 # Start app
-print("Vigilant bot has been started...")
-fl = Process(target=run_flask)
-fl.start()
-import run
+if __name__ == "__main__":
+    print(f"WebApp is running on port {Environment().WEB_APP_PORT}")
+    fl = Process(target=run_flask)
+    fl.start()
+    import run